Showing dashboards in OO grafana
Moving the dashboards from the old monitoring stack to the new
