+++ pkg/products/monitoringcommon/dashboards/resourceByNamespace.go
@@ -1,4 +1,4 @@
-package monitoring
+package monitoringcommon
 
 import (
 	"github.com/integr8ly/integreatly-operator/apis/v1alpha1"
@@ -7,7 +7,7 @@ import (
 
 // This dashboard json is dynamically configured based on installation type (rhmi or rhoam)
 // The installation name taken from the v1alpha1.RHMI.ObjectMeta.Name
-func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName string) string {
+func GetMonitoringGrafanaDBResourceByNSJSON(nsPrefix, installationName string) string {
 	quota := ``
 	if installationName == resources.InstallationNames[string(v1alpha1.InstallationTypeManagedApi)] || installationName == resources.InstallationNames[string(v1alpha1.InstallationTypeMultitenantManagedApi)] {
 		quota = `,
@@ -111,30 +111,14 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 			"stack": false,
 			"steppedLine": false,
 			"targets": [{
-					"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace', pod=~'$pod', container!='POD'}) by (pod)",
-					"format": "time_series",
-					"intervalFactor": 2,
-					"legendFormat": "{{pod}}",
-					"legendLink": null,
-					"step": 10
-				},
-				{
-					"expr": "kube_pod_container_resource_requests{namespace=~'$namespace', pod=~'$pod',resource='cpu'}",
-					"format": "time_series",
-					"intervalFactor": 2,
-					"legendFormat": "{{pod}}",
-					"legendLink": null,
-					"step": 10
-				},
-				{
-					"expr": "kube_pod_container_resource_limits{namespace=~'$namespace', pod=~'$pod', resource='cpu'}",
-					"format": "time_series",
-					"intervalFactor": 2,
-					"legendFormat": "{{pod}} Limit",
-					"legendLink": null,
-					"step": 10
-				}
-			],
+				"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace'}) by (pod)",
+				"format": "time_series",
+				"intervalFactor": 2,
+				"legendFormat": "{{pod}}",
+				"legendLink": null,
+				"step": 10,
+				"refId": "A"
+			}],
 			"thresholds": [],
 			"timeFrom": null,
 			"timeRegions": [],
@@ -227,8 +211,8 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 			"seriesOverrides": [],
 			"showHeader": true,
 			"sort": {
-				"col": 0,
-				"desc": true
+				"col": 1,
+				"desc": false
 			},
 			"spaceLength": 10,
 			"stack": false,
@@ -310,15 +294,15 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"unit": "percentunit"
 				},
 				{
-					"alias": "Container",
+					"alias": "Pod",
 					"colorMode": null,
 					"colors": [],
 					"dateFormat": "YYYY-MM-DD HH:mm:ss",
 					"decimals": 2,
-					"link": false,
+					"link": true,
 					"linkTooltip": "Drill down",
-					"linkUrl": "",
-					"pattern": "container",
+					"linkUrl": "/d/c84ae905b9f54268be6be82c9a5b7dd6/resources-by-pod?var-namespace=$namespace&var-pod=$__cell",
+					"pattern": "pod",
 					"thresholds": [],
 					"type": "number",
 					"unit": "short"
@@ -336,7 +320,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 				}
 			],
 			"targets": [{
-					"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace', pod=~'$pod', container!='POD'}) by (container)",
+					"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -345,7 +329,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(kube_pod_container_resource_requests{namespace=~'$namespace', pod=~'$pod',resource='cpu'}) by (container)",
+					"expr": "sum(kube_pod_container_resource_requests{namespace=~'$namespace',resource='cpu'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -354,7 +338,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace', pod=~'$pod'}) by (container) / sum(kube_pod_container_resource_requests{namespace=~'$namespace', pod=~'$pod',resource='cpu'}) by (container)",
+					"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace'}) by (pod) / sum(kube_pod_container_resource_requests{namespace=~'$namespace',resource='cpu'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -363,7 +347,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(kube_pod_container_resource_limits{namespace=~'$namespace', pod=~'$pod', resource='cpu'}) by (container)",
+					"expr": "sum(kube_pod_container_resource_limits{namespace=~'$namespace',resource='cpu'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -372,7 +356,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace', pod=~'$pod'}) by (container) / sum(kube_pod_container_resource_limits{namespace=~'$namespace', pod=~'$pod', resource='cpu'}) by (container)",
+					"expr": "sum(node_namespace_pod_container:container_cpu_usage_seconds_total:sum_rate{namespace=~'$namespace'}) by (pod) / sum(kube_pod_container_resource_limits{namespace=~'$namespace',resource='cpu'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -467,30 +451,14 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 			"stack": false,
 			"steppedLine": false,
 			"targets": [{
-					"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace', pod=~'$pod', container !='',container !='POD'}) by (pod)",
-					"format": "time_series",
-					"intervalFactor": 2,
-					"legendFormat": "{{pod}}",
-					"legendLink": null,
-					"step": 10
-				},
-				{
-					"expr": "kube_node_status_allocatable{namespace=~'$namespace', pod=~'$pod', resource='memory'}",
-					"format": "time_series",
-					"intervalFactor": 2,
-					"legendFormat": "{{pod}} Request",
-					"legendLink": null,
-					"step": 10
-				},
-				{
-					"expr": "kube_pod_container_resource_limits{namespace=~'$namespace', pod=~'$pod',resource='memory'}",
-					"format": "time_series",
-					"intervalFactor": 2,
-					"legendFormat": "{{pod}} Limit",
-					"legendLink": null,
-					"step": 10
-				}
-			],
+				"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace', container=''}) by (pod)",
+				"format": "time_series",
+				"intervalFactor": 2,
+				"legendFormat": "{{pod}}",
+				"legendLink": null,
+				"step": 10,
+				"refId": "A"
+			}],
 			"thresholds": [],
 			"timeFrom": null,
 			"timeRegions": [],
@@ -583,7 +551,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 			"seriesOverrides": [],
 			"showHeader": true,
 			"sort": {
-				"col": 0,
+				"col": 1,
 				"desc": true
 			},
 			"spaceLength": 10,
@@ -666,15 +634,15 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"unit": "percentunit"
 				},
 				{
-					"alias": "Container",
+					"alias": "Pod",
 					"colorMode": null,
 					"colors": [],
 					"dateFormat": "YYYY-MM-DD HH:mm:ss",
 					"decimals": 2,
-					"link": false,
+					"link": true,
 					"linkTooltip": "Drill down",
-					"linkUrl": "",
-					"pattern": "container",
+					"linkUrl": "/d/c84ae905b9f54268be6be82c9a5b7dd6/resources-by-pod?var-namespace=$namespace&var-pod=$__cell",
+					"pattern": "pod",
 					"thresholds": [],
 					"type": "number",
 					"unit": "short"
@@ -692,7 +660,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 				}
 			],
 			"targets": [{
-					"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace', pod=~'$pod', container !=''}) by (container)",
+					"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace',container=''}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -701,7 +669,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(kube_pod_container_resource_requests{namespace=~'$namespace', pod=~'$pod',resource='memory'}) by (container)",
+					"expr": "sum(kube_pod_container_resource_requests{namespace=~'$namespace',resource='memory'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -710,7 +678,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace', pod=~'$pod'}) by (container) / sum(kube_pod_container_resource_requests{namespace=~'$namespace', pod=~'$pod',resource='memory'}) by (container)",
+					"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace',container=''}) by (pod) / sum(kube_pod_container_resource_requests{namespace=~'$namespace',resource='memory'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -719,7 +687,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(kube_pod_container_resource_requests{namespace=~'$namespace', pod=~'$pod', resource='memory'}) by (container)",
+					"expr": "sum(kube_pod_container_resource_limits{namespace=~'$namespace',resource='memory'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -728,7 +696,7 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 					"step": 10
 				},
 				{
-					"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace', pod=~'$pod', container!=''}) by (container) / sum(kube_pod_container_resource_requests{namespace=~'$namespace', pod=~'$pod', resource='memory'}) by (container)",
+					"expr": "sum(container_memory_working_set_bytes{namespace=~'$namespace',container=''}) by (pod) / sum(kube_pod_container_resource_limits{namespace=~'$namespace',resource='memory'}) by (pod)",
 					"format": "table",
 					"instant": true,
 					"intervalFactor": 2,
@@ -780,48 +748,26 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 	"tags": [],
 	"templating": {
 		"list": [{
-				"allValue": null,
-				"datasource": "Prometheus",
-				"definition": "",
-				"hide": 0,
-				"includeAll": false,
-				"label": "namespace",
-				"multi": false,
-				"name": "namespace",
-				"options": [],
-				"query": "query_result(count(kube_namespace_labels{namespace=~'` + namespacePrefix + `.*'}) by (namespace))",
-				"refresh": 1,
-				"regex": "/\"(.*?)\"/",
-				"skipUrlSync": false,
-				"sort": 1,
-				"tagValuesQuery": "",
-				"tags": [],
-				"tagsQuery": "",
-				"type": "query",
-				"useTags": false
-			},
-			{
-				"allValue": null,
-				"datasource": "Prometheus",
-				"definition": "",
-				"hide": 0,
-				"includeAll": false,
-				"label": "pod",
-				"multi": false,
-				"name": "pod",
-				"options": [],
-				"query": "label_values(kube_pod_info{namespace=~'$namespace'}, pod)",
-				"refresh": 1,
-				"regex": "",
-				"skipUrlSync": false,
-				"sort": 1,
-				"tagValuesQuery": "",
-				"tags": [],
-				"tagsQuery": "",
-				"type": "query",
-				"useTags": false
-			}
-		]
+			"allValue": null,
+			"datasource": "Prometheus",
+			"definition": "",
+			"hide": 0,
+			"includeAll": false,
+			"label": "namespace",
+			"multi": false,
+			"name": "namespace",
+			"options": [],
+			"query": "query_result(count(kube_namespace_labels{namespace=~'` + nsPrefix + `.*'}) by (namespace))",
+			"refresh": 1,
+			"regex": "/\"(.*?)\"/",
+			"skipUrlSync": false,
+			"sort": 1,
+			"tagValuesQuery": "",
+			"tags": [],
+			"tagsQuery": "",
+			"type": "query",
+			"useTags": false
+		}]
 	},
 	"time": {
 		"from": "now-1h",
@@ -853,8 +799,8 @@ func GetMonitoringGrafanaDBResourceByPodJSON(namespacePrefix, installationName s
 		]
 	},
 	"timezone": "",
-	"title": "Resource Usage By Pod",
-	"uid": "c84ae905b9f54268be6be82c9a5b7dd6",
+	"uid": "a9ce5290ba1d485ca67e05c0a63aa2d8",
+	"title": "Resource Usage By Namespace",
 	"version": 2
 }`
 }
